refactor: 重构 VideoSlim.py
