--- VideoSlim/app/__init__.py
@@ -1,3 +1,0 @@
-import config
-import core
-import log
