+++ VideoSlim/src/__init__.py
@@ -1,0 +1,3 @@
+from .config import *
+from .controller import *
+from .view import *
